Add makefile to package
Creates a fully working .dll by running the 'make' command. Mostly
tailored to work with Linux, could be used on other systems.

Modified the gitignore to include the package/ directory created during
the package process, as well as *.swp files that are used by vim as swap
files.

This includes the LICENSE.txt as well as the CHANGES.txt from the
Toolbar directory.